plaid-sync/requirements.txt
plaid-python==7.1.0
pandas
influxdb-client
notion-client
